adding shared test logic and creating new test fixture for maui tests
